gpu: nvgpu: Fix signed comparison bugs
Fix small problems related to signed versus unsigned comparisons
throughout the driver. Bump up the warning level to prevent such
problems from occuring in future.

Reviewed-on: http://git-master/r/1252068
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/gk20a/channel_sync_gk20a.c
@@ -192,7 +192,7 @@ static int __gk20a_channel_syncpt_incr(struct gk20a_channel_sync *s,
 				       bool need_sync_fence)
 {
 	u32 thresh;
-	int incr_cmd_size;
+	size_t incr_cmd_size;
 	int off;
 	int err;
 	struct gk20a_channel_syncpt *sp =
